requirements.txt
appdirs==1.4.4
astroid==2.6.4
attrs==21.2.0
black==22.1.0
boto3==1.18.22
botocore==1.21.22
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
docopt==0.6.2
eight==1.0.1
Flask==2.0.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.1
idna==3.2
isodate==0.6.0
isort==5.9.2
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
lazy-object-proxy==1.6.0
lxml==4.6.3
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
mccabe==0.6.1
mypy-extensions==0.4.3
num2words==0.5.10
openpyxl==3.1.2
pathspec==0.9.0
pdfkit==0.6.1
platformdirs==2.4.0
pycparser==2.20
pydantic==1.8.2
pylint==2.9.4
PyMySQL==1.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
qrcode==7.3
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==0.9.1
s3transfer==0.5.0
signxml==2.8.2
six==1.16.0
SQLAlchemy==1.4.22
suds-community==0.8.5
toml==0.10.2
typing_extensions==4.0.0
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
zeep==4.1.0
pypdf==3.1.0
PyPDF2==3.0.1
Pillow==8.3.2
